build(icons): TRAC-1519 rewrite downloader script off inquirer 8
`inquirer` went ESM-only at v9 and was rearchitected around
`@inquirer/prompts`, breaking the `require()` + `inquirer.registerPrompt`
pattern `downloader.js` relied on. Rename to `downloader.mjs` and rewrite
as a native ES module (static imports, top-level await,
`import.meta.dirname`).

Bump `inquirer` 8.2.7 -> 14.0.2. Drop `inquirer-autocomplete-prompt`
entirely rather than bump it to 3.0.1 as originally scoped: that version
still peer-deps on `inquirer@^9` and deep-imports internals
(`inquirer/lib/prompts/base.js`, etc.) removed from inquirer 14's exports
map, confirmed by actually running the script (`ERR_PACKAGE_PATH_NOT_EXPORTED`).
Replaced with inquirer 14's built-in `search` prompt, which provides the
same async-source, type-to-filter UX with no extra dependency.

`lint` now passes `--ext .js,.mjs` so the new file is covered by ESLint
(previously silently skipped under the old default extensions). Removed
the now-stale dependabot major-version ignore rules for both packages.

Dev-only script, no consumer or build-output impact.
