+++ .github/dependabot.yml
@@ -5,9 +5,3 @@ updates:
     schedule:
       interval: 'weekly'
       day: 'tuesday'
-    ignore:
-      # Next major version is exported as an ESM module
-      - dependency-name: 'inquirer'
-        update-types: ['version-update:semver-major']
-      - dependency-name: 'inquirer-autocomplete-prompt'
-        update-types: ['version-update:semver-major']
